Split cli so that we only build app in there
