+++ src/main.rs
@@ -39,6 +39,38 @@ fn main() {
             &args
         );
     } else {
-        cli::run();
+        let matches = cli::make_app().get_matches();
+
+        match matches.subcommand() {
+            ("exec", Some(matches)) => {
+                let args = matches.values_of("command_args")
+                    .unwrap_or_default()
+                    .map(|i| i.to_owned())
+                    .collect::<Vec<String>>();
+
+                exec_cmd::run(
+                    matches.value_of("command").unwrap(),
+                    &args
+                )
+            },
+            ("which", Some(matches)) =>
+                which_cmd::run(matches.value_of("command").unwrap()),
+            ("shim", Some(_)) => shim_cmd::run(),
+            ("scan", Some(matches)) =>
+                scan_cmd::run(matches.value_of("command").unwrap()),
+            ("use", Some(matches)) =>
+                use_cmd::run(
+                    matches.value_of("command").unwrap(),
+                    matches.value_of("version")
+                ),
+            ("show", Some(_)) => show_cmd::run(),
+            ("def", Some(matches)) =>
+                def_cmd::run(
+                    matches.value_of("command").unwrap(),
+                    matches.value_of("version").unwrap(),
+                    matches.value_of("bin").unwrap()
+                ),
+            _ => unreachable!(),
+        };
     }
 }
